+++ tests/test_specials.py
@@ -1,5 +1,5 @@
-  #!/usr/bin/python
-  # -*- coding: utf-8 -*-
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
 
 from dbutils import dbtest, POSTGRES_USER, foreign_db_environ
 import itertools