+++ changelog.rst
@@ -5,6 +5,7 @@ Features:
 ---------
 
 * Add support for the `\dp` metacommand that lists the privileges of Postgres objects (Thanks: `Guru Devanla`_)
+* Add support for the `\np <named_query_pattern>` metacommand that returns named queries matching the pattern (Thanks: `Guru Devanla`_)
 
 1.11.10
 ======